refactor: Extract domain exceptions into domain/exceptions.py
Separate exception classes (SystemException, DomainException and subclasses)
from error data classes (Error) so each concept has its own module.
