+++ sharedkernel/domain/__init__.py
@@ -1,9 +1,10 @@
-from . import data, errors, events, models, repositories, services
+from . import data, errors, events, exceptions, models, repositories, services
 
 __all__ = [
     "data",
     "errors",
     "events",
+    "exceptions",
     "models",
     "repositories",
     "services",
