+++ sharedkernel/domain/models.py
@@ -2,7 +2,7 @@ from collections import deque
 from dataclasses import dataclass
 from uuid import UUID
 
-from sharedkernel.domain.errors import UnhandledEventType
+from sharedkernel.domain.exceptions import UnhandledEventType
 from sharedkernel.domain.events import DomainEvent
